+++ chef-repo/cookbooks/docker/metadata.rb
@@ -4,7 +4,7 @@ maintainer_email  'bflad417@gmail.com'
 license           'Apache 2.0'
 description       'Installs/Configures Docker'
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version           '0.20.0'
+version           '0.24.2'
 recipe            'docker', 'Installs/Configures Docker'
 recipe            'docker::aufs', 'Installs/Loads AUFS Linux module'
 recipe            'docker::binary', 'Installs Docker binary'
@@ -16,10 +16,17 @@ recipe            'docker::systemd', 'Installs/Starts Docker via systemd'
 recipe            'docker::sysv', 'Installs/Starts Docker via SysV'
 recipe            'docker::upstart', 'Installs/Starts Docker via Upstart'
 
-%w{ centos fedora oracle redhat ubuntu }.each do |os|
-  supports os
-end
+supports 'centos', '>= 6.0'
+supports 'debian', '>= 7.0'
+supports 'fedora', '>= 19.0'
+supports 'oracle', '>= 6.0'
+supports 'redhat', '>= 6.0'
+supports 'ubuntu', '>= 12.04'
 
-%w{ apt git golang lxc modules yum }.each do |cb|
-  depends cb
-end
+depends 'apt'
+depends 'git'
+depends 'golang'
+depends 'lxc', '>= 1.1.6'
+depends 'modules'
+depends 'sysctl'
+depends 'yum-epel'