+++ chef-repo/cookbooks/nginx-site/metadata.rb
@@ -1,4 +1,0 @@
-maintainer        "Yuanying"
-maintainer_email  "yuanying@3machinae.com"
-description       "Configures Nginx Sites"
-version           "0.1"